+++ test/libsolidity/semanticTests/externalContracts/prbmath_signed.sol
@@ -48,8 +48,8 @@ contract test {
 }
 // ----
 // constructor()
-// gas irOptimized: 177903
-// gas irOptimized code: 1674400
+// gas irOptimized: 180236
+// gas irOptimized code: 1717200
 // gas legacy: 209723
 // gas legacy code: 2205000
 // gas legacyOptimized: 178012
@@ -57,42 +57,42 @@ contract test {
 // gas ssaCFGOptimized: 175028
 // gas ssaCFGOptimized code: 1639200
 // div(int256,int256): 3141592653589793238, 88714123 -> 35412542528203691288251815328
-// gas irOptimized: 22045
+// gas irOptimized: 22067
 // gas legacy: 22736
 // gas legacyOptimized: 22264
 // exp(int256): 3141592653589793238 -> 23140692632779268978
-// gas irOptimized: 24452
+// gas irOptimized: 24364
 // gas legacy: 25124
 // gas legacyOptimized: 24351
 // exp2(int256): 3141592653589793238 -> 8824977827076287620
-// gas irOptimized: 24162
+// gas irOptimized: 24075
 // gas legacy: 24787
 // gas legacyOptimized: 24105
 // gm(int256,int256): 3141592653589793238, 88714123 -> 16694419339601
-// gas irOptimized: 22805
+// gas irOptimized: 22718
 // gas legacy: 23202
 // gas legacyOptimized: 22685
 // log10(int256): 3141592653589793238 -> 4971498726941338506
-// gas irOptimized: 30300
+// gas irOptimized: 28819
 // gas legacy: 32841
 // gas legacyOptimized: 30249
 // log2(int256): 3141592653589793238 -> 1651496129472318782
-// gas irOptimized: 28566
+// gas irOptimized: 28479
 // gas legacy: 30979
 // gas legacyOptimized: 28357
 // mul(int256,int256): 3141592653589793238, 88714123 -> 278703637
-// gas irOptimized: 22147
+// gas irOptimized: 22060
 // gas legacy: 22775
 // gas legacyOptimized: 22288
 // pow(int256,uint256): 3141592653589793238, 5 -> 306019684785281453040
-// gas irOptimized: 22488
+// gas irOptimized: 22510
 // gas legacy: 23453
 // gas legacyOptimized: 22921
 // sqrt(int256): 3141592653589793238 -> 1772453850905516027
-// gas irOptimized: 22458
+// gas irOptimized: 22480
 // gas legacy: 22784
 // gas legacyOptimized: 22420
 // benchmark(int256): 3141592653589793238 -> 998882724338592125, 1000000000000000000, 1000000000000000000
-// gas irOptimized: 34893
+// gas irOptimized: 34915
 // gas legacy: 35244
 // gas legacyOptimized: 33996