+++ test/libsolidity/semanticTests/storage/storage_boundary_struct_array_multislot.sol
@@ -92,7 +92,7 @@ contract C {
 // boundaryArray() -> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
 // destArray() -> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
 // fillBoundaryArray()
-// gas irOptimized: 688719
+// gas irOptimized: 688654
 // gas legacy: 700075
 // gas legacyOptimized: 691605
 // gas ssaCFGOptimized: 688794
@@ -100,7 +100,7 @@ contract C {
 // boundaryArray() -> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30
 // destArray() -> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0
 // copyFromBoundary()
-// gas irOptimized: 748779
+// gas irOptimized: 748801
 // gas legacy: 767297
 // gas legacyOptimized: 748756
 // gas ssaCFGOptimized: 748798
@@ -108,7 +108,7 @@ contract C {
 // boundaryArray() -> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30
 // destArray() -> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30
 // fillDestArray()
-// gas irOptimized: 175623
+// gas irOptimized: 175642
 // gas legacy: 187093
 // gas legacyOptimized: 178625
 // gas ssaCFGOptimized: 175698
@@ -116,7 +116,7 @@ contract C {
 // boundaryArray() -> 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30
 // destArray() -> 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60
 // copyToBoundary()
-// gas irOptimized: 235823
+// gas irOptimized: 235758
 // gas legacy: 254252
 // gas legacyOptimized: 235722
 // gas ssaCFGOptimized: 235842
@@ -124,7 +124,7 @@ contract C {
 // boundaryArray() -> 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60
 // destArray() -> 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60
 // deleteBoundaryArray()
-// gas irOptimized: 137824
+// gas irOptimized: 137771
 // gas legacy: 137971
 // gas legacyOptimized: 137750
 // gas ssaCFGOptimized: 137819